+++ Street Fighter II/Street-Fighter-II/ModulePlayer2.h
@@ -66,7 +66,7 @@ public:
 	bool Bcrouch = false;
 	bool Bstand = false;
 
-	int life = 100;
+	float life = 100;
 
 	/*Mix_Chunk* kickfx = nullptr;
 	Mix_Chunk* punchfx = nullptr;*/
